Finalmente os testes funcionaram

+++ controllers/TaskController.js
@@ -2,20 +2,25 @@ const Task = require("../models/Task");
 const User = require("../models/User");
 
 module.exports = {
+  //Criar tarefas
   async createTask(req, res) {
     try {
       const user = await User.findById(req.userId);
       const task = await Task.create({ ...req.body, user: req.userId });
       user.tasks.push(task);
       await user.save();
-      return res.send(task);
+      return res.send({
+        name: req.body.name,
+        priority: req.body.priority,
+        description: req.body.description,
+      });
     } catch (error) {
       return res
         .status(400)
         .send("Erro na criação, confira se preencheu os dados corretamente");
     }
   },
-
+  //Listar tarefas
   async listTasks(req, res) {
     try {
       const user = await User.findById(req.userId);
@@ -24,7 +29,7 @@ module.exports = {
       return res.status(400).send(error);
     }
   },
-
+  //Ordenar listas pela prioridade
   async sortTasks(req, res) {
     try {
       const user = await User.findById(req.userId);
@@ -41,7 +46,7 @@ module.exports = {
       return res.status(400).send(error);
     }
   },
-
+  //Editar tarefa
   async updateTask(req, res) {
     try {
       const user = await User.findById(req.userId);
@@ -72,30 +77,35 @@ module.exports = {
       return res.status(400).send(error);
     }
   },
-
+  //Deletar tarefa
   async deleteTask(req, res) {
     try {
       const user = await User.findById(req.userId);
+      let flag = false;
       user.tasks.forEach((task) => {
         if (task._id == req.params.taskId) {
           let index = user.tasks.indexOf(task);
           user.tasks.splice(index, 1);
+          flag = true;
         }
       });
-      const taskToDelete = await Task.deleteOne({ _id: req.params.taskId });
-      await user.save();
-      return res.send(user.tasks);
+      if (flag) {
+        const taskToDelete = await Task.deleteOne({ _id: req.params.taskId });
+        await user.save();
+        return res.send(user.tasks);
+      }
+      return res.status(400).send("Tarefa não encontrada.");
     } catch (error) {
       console.log(error);
-      return res.status(400).send("Erro ao deletar a tarefa");
+      return res.status(400).send("Erro ao deletar a tarefa.");
     }
   },
-
-  async deleteData(req, res) {
-    await Task.deleteMany();
-    const user = await User.findById(req.userId);
-    user.tasks.splice(0, user.tasks.length);
-    await user.save();
-    return res.status(200).send("Deletado com sucesso");
-  },
+  // Deletar todas as tarefas e limpar lista de tarefas do usuário.
+  // async deleteData(req, res) {
+  //   await Task.deleteMany();
+  //   const user = await User.findById(req.userId);
+  //   user.tasks.splice(0, user.tasks.length);
+  //   await user.save();
+  //   return res.status(200).send("Deletado com sucesso");
+  // },
 };
